+++ docs/ModFramework-Example-Mod/changelog.md
@@ -43,7 +43,10 @@
     - Added example for getting a modded solenoid reference
     - Added example for using framework types to reference base solenoids
 - ### Pilot
+    - Added example for adding a modded pilot template
     - Added example for adding pilots on game start
+    - Added example for getting a modded pilot template reference
+    - Added example for adding pilot based on a modded template
 - ### Research
     - Added example for adding a modded research
     - Added example for getting a modded research reference